requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2026.2
uritemplate==4.2.0
